requirements.txt
discord.py
nop
pyyaml
mysqlclient
pymysql
sqlalchemy
alembic
asyncio-mqtt
